Move get_container_id_for_mesos_id into mesos_tools
